requirements.txt
fastapi>=0.109.0
argon2-cffi>=23.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.27.0
